Create exercise1&2.py
Week 1 Discussion Board:
My name is Aideen, so the first and last letter of my name (A + N = 1 + 14) give the number 15.  The 15th Fibonacci number is 610.

Week 2 Discussion Board:
My surname is Byrne
The first letter B is number 66
The last letter e is number 101
Fibonacci number 167 is 35600075545958458963222876581316753
Ord() used in a string returns an integer representing the Unicode code point of a character e.g. ord (' a ') returns the integer 97.